requirements.txt
flask==2.3.3
PyPDF2==3.0.1
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
openai==1.12.0
httpx>=0.23.0
tqdm==4.66.1
pdf2image==1.17.0